fix colors & refractor
- rangement code dans fichier home_screen
- clean main
- fix light/dark theme
